add -m argument to "beancmd flush"

+++ beancmd/flush.py
@@ -12,6 +12,8 @@ def setup_parser(parser=None):
     parser.add_argument('-H', '--host', default='localhost', help='Host of beanstalk server (default %(default)s)')
     parser.add_argument('-p', '--port', default=11300, type=int, help='Port of beanstalk server (default %(default)s)')
     parser.add_argument('-y', '--yes', action='store_true', help='Do not prompt when deleting things')
+    parser.add_argument('-m', '--mode', action='append', default=[], choices=('ready', 'delayed', 'buried'),
+                        help='What kind of jobs to purge (may be repeated; if not passed, will purge all)')
     parser.add_argument('tubes', nargs='*', help='Tubes to flush (if not passed, does them all!)')
     return parser
 
@@ -28,24 +30,35 @@ def run(args):
     delayed_jobs = 0
     buried_jobs = 0
 
+    if args.mode:
+        mode = set(args.mode)
+    else:
+        mode = set(['ready', 'delayed', 'buried'])
+
     for tube in tubes:
         stats = client.stats_tube(tube)
-        ready_jobs += stats['current-jobs-ready']
-        delayed_jobs += stats['current-jobs-delayed']
-        buried_jobs += stats['current-jobs-buried']
+        if 'ready' in mode:
+            ready_jobs += stats['current-jobs-ready']
+        if 'delayed' in mode:
+            delayed_jobs += stats['current-jobs-delayed']
+        if 'buried' in mode:
+            buried_jobs += stats['current-jobs-buried']
 
     client.watch('unused-fake-tube')
     for tube in tubes:
         client.watch(tube)
         client.use(tube)
-        print('Flushing {0} READY jobs'.format(ready_jobs))
-        for job in util.progress(client.reserve_iter(), total=ready_jobs, unit='jobs'):
-            client.delete_job(job.job_id)
-        print('Flushing {0} DELAYED jobs'.format(delayed_jobs))
-        for job in util.progress(client.peek_delayed_iter(), total=delayed_jobs, unit='jobs'):
-            client.delete_job(job.job_id)
-        print('Flushing {0} BURIED jobs'.format(buried_jobs))
-        for job in util.progress(client.peek_buried_iter(), total=buried_jobs, unit='jobs'):
-            client.delete_job(job.job_id)
+        if 'ready' in mode:
+            print('Flushing {0} READY jobs'.format(ready_jobs))
+            for job in util.progress(client.reserve_iter(), total=ready_jobs, unit='jobs'):
+                client.delete_job(job.job_id)
+        if 'delayed' in mode:
+            print('Flushing {0} DELAYED jobs'.format(delayed_jobs))
+            for job in util.progress(client.peek_delayed_iter(), total=delayed_jobs, unit='jobs'):
+                client.delete_job(job.job_id)
+        if 'buried' in mode:
+            print('Flushing {0} BURIED jobs'.format(buried_jobs))
+            for job in util.progress(client.peek_buried_iter(), total=buried_jobs, unit='jobs'):
+                client.delete_job(job.job_id)
         client.ignore(tube)
     return 0
